New method for obtaining attachment's filename
Added the parameter "filenameFromAttachmentName" which can be "Y" or "N"
(default). If "Y" the filename for the new messages will be derived from
the attachment's name field from the original message.

This may be used, if the attachment is a binary file. (In this case the
original retrieveFilename()-method is useless.)