New feature - only allow valid characters in XML element name
New parameter 'onlyValidCharsInXMLName' is considered, when parameter
'processFieldNames' = 'fromFile'. 

If the parameter is equal to 'N' (default) XML element names are handled
the same way as before the new feature was included.

If the parameter is equal to 'Y' each header column name is checked for
invalid characters and they are stripped from the XML element name.