New feature - Header row is now configurable.
Two new parameters - 'headerRow' and 'rowOffsetFromHeaderRow' are now
considered, when parameter 'processFieldNames' = 'fromFile'.

If the parameter 'headerRow' is set to a positive integer value or 0
(default) the XML element names are created from that row.

If parameter 'rowOffsetFromHeaderRow' is equal to 'N' (default)
functionality will continue as before introduction of the new feature.

If the parameter 'rowOffsetFromHeaderRow' is equal to 'Y' the value in
parameter 'rowOffset' is added to the value in parameter 'headerRow'. If
parameter 'rowOffset' is set to '0' (default) it will automatically be
adjusted to '1', before adding it to the parameter 'headerRow'.